feat: bind counry page to server
