+++ src/assets/locales/en/translation.json
@@ -101,6 +101,3 @@
     "learn_more": "Learn more"
   }
 }
-
-
-
