--- src/data.js
@@ -1,206 +1,0 @@
-const data = {
-  brazil: {
-    name: 'Brazil',
-    imageUrl: 'https://pix10.agoda.net/geo/country/207/3_207_brazil_02.jpg?s=1920x',
-    videoUrl: '5STHnV-ZRfE',
-    galery: [
-      {
-        img: 'https://top10.travel/wp-content/uploads/2016/05/iguasu.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2016/05/statuya-khrista-iskupitelya.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2016/05/pantanal.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2016/05/salvador-baiya.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2016/05/vodopad-karakol.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2016/05/ipanema.jpg',
-      },
-    ],
-  },
-  italy: {
-    name: 'Italy',
-    imageUrl: 'https://www.atlanticcouncil.org/wp-content/uploads/2020/09/Rome-coroavirus-large-1024x683.jpg',
-    videoUrl: 'FlRwssZYRM0',
-    galery: [
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/10/grand-canal-veneciya.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/10/kolizey.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/10/casteldelmonte.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/10/pizanskaya-bashnya.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/10/milansky-sobor.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/10/fontan-trevi.jpg',
-      },
-    ],
-  },
-  spain: {
-    name: 'Spain',
-    imageUrl: 'https://abstour.by/upload/iblock/39b/39b113eee670389355bc92b6be574b44.jpg',
-    videoUrl: 'clnt7kk6rIU',
-    galery: [
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/hram-sviatogo-semeistva.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/dom-mila-balo.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/goticheskiy-kvartal.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/ploshhad-ispanii.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/park-guyelya.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/alkasar-v-segovii.jpg',
-      },
-    ],
-  },
-  usa: {
-    name: 'USA',
-    imageUrl: 'https://aif-s3.aif.ru/images/021/486/ee42297d09cbe676f8580c86b2d603c2.jpg',
-    videoUrl: 'yRJFLKm7-Nk',
-    galery: [
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/statuya-svobody.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/rushmore.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/most-zolotye-vorota.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/times-square.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/grand-canyon.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/kamni-v-doline-smerti.jpg',
-      },
-    ],
-  },
-  australia: {
-    name: 'Australia',
-    imageUrl: 'https://kipmu.ru/wp-content/uploads/avstrlya-696x440.jpg',
-    videoUrl: 'zJvCrUlzSvc',
-    galery: [
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/12/bolshoy-barierniy-rif.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/12/sydney_opera_house_sails.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/12/zdanie-korolevy-viktorii.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/12/muzeya-mona.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/12/katorzhnoe-poselenie-port-artur.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/12/korolevskiy-sad-sidneya.jpg',
-      },
-    ],
-  },
-  england: {
-    name: 'England',
-    imageUrl:
-      'https://n1s1.hsmedia.ru/0d/c3/bc/0dc3bc2381a2e155424ca7a682419ada/620x462_1_94ce4a400cea100bb26d548921ee012f@1000x745_0xac120003_13763038881563803374.jpg',
-    videoUrl: 'ZS2v51y8TiU',
-    galery: [
-      {
-        img: 'https://top10.travel/wp-content/uploads/2015/05/tauerskij-most-1.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2015/05/stounkhendzh.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2015/05/londonskiy-glaz.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2015/05/bashnya-elizavety.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2015/05/vindzorskij-zamok.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2015/05/uollaton-holl.jpg',
-      },
-    ],
-  },
-  irland: {
-    name: 'Irland',
-    imageUrl:
-      'https://otdyhateli.com/wp-content/uploads/2016/03/ci-travel-channel-ashford-castle-ireland.jpg.rend_.hgtvcom.966.725.jpeg',
-    videoUrl: 'u-UA8t2EVpA',
-    galery: [
-      {
-        img: 'https://top10.travel/wp-content/uploads/2017/02/kellskaya-kniga.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2017/02/dublinskiy-zamok.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2017/02/sobor-svyatogo-patrika.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2017/02/nacionalnyj-muzej-irlandii.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2017/02/tyurma-kilmyenhem.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2017/02/most-semyuela-bekketa.jpg',
-      },
-    ],
-  },
-  france: {
-    name: 'France',
-    imageUrl: 'https://www.biletik.aero/upload/resize_cache/format_converted/0174a292e2ce40327282349518386ec2.webp',
-    videoUrl: '8f2tVaHQumM',
-    galery: [
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/eyfeleva-bashnya.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/triumfalnaya-arka-1.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/luvr.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/disneyland-paris.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/versalskiy-dvorec.jpg',
-      },
-      {
-        img: 'https://top10.travel/wp-content/uploads/2014/09/sad-kloda-mone.jpg',
-      },
-    ],
-  },
-};
-
-export default data;
